requirements.txt
fastapi  
uvicorn[standard]  
google-genai
requests
beautifulsoup4